The client benchmark measures the schema it says it mirrors
Its module doc says every shape is byte-identical to the authoritative
DDL. Three executable definitions still created UInt64 fingerprint
columns, the row structs carried u64, and the payload accounting charged
8 bytes for a 16-byte column, so it measured a narrower schema than the
one that ships and undercounted every byte figure it reported.

  the two table DDLs and the tier DDL   UInt128
  the row structs and both transports   u128, in the generators, the
                                        report types and both candidate
                                        crates' row types
  payload_bytes                         16 bytes for the fingerprint
  the generators                        both 64-bit words are filled, from
                                        two independent draws of the mix
  the aggregate-state predicate         minted, not a bare decimal: the
                                        value is above 2^64 and a bare
                                        decimal there is read as Float64

Re-run against ClickHouse 26.3.29.7, small scale, on the shared server:

  aggstate   both crates ok=true, fingerprint
             340282366920938463205120190399834488818 exact, val_count
             1000, first_value 1.0, last_value 1000.0
  ddl        both crates reliable=true -- CREATE TABLE, the materialized
             view and the chunked backfill over the widened column with
             CODEC(Delta(8), ZSTD(1))
  insert     200,000 rows, 50,000 per block, 3 reps: metric rows
             2,580,869 and 1,578,397 rows/s at p50, log rows 731,262 and
             752,286
  fetch      400 rows, p50 4.4ms and 3.9ms, both candidates agreeing on
             the checksum

Those figures are from this machine at a fraction of the spike's row
count and are not comparable with the record's; they are evidence that
the widened path runs on both transports, not a new decision.

A new test asserts what the generators now produce: no row has an empty
high or low word, and the values stay distinct. The constant sets the top
bit of both words, with the two words different so a transport that
swapped them would be visible.

The decision record and the architecture table now say which width each
figure was taken at, and record the re-measurement.

+++ xtask/src/ch_bench/ddl.rs
@@ -31,7 +31,7 @@ pub fn tier_table_ddl(tier_table: &str) -> String {
     format!(
         "CREATE TABLE IF NOT EXISTS {tier_table} (
             metric_name   LowCardinality(String),
-            fingerprint   UInt64                                 CODEC(Delta(8), ZSTD(1)),
+            fingerprint   UInt128                                 CODEC(Delta(8), ZSTD(1)),
             ts            DateTime                               CODEC(DoubleDelta, ZSTD(1)),
             val_min       SimpleAggregateFunction(min, Float64)  CODEC(Gorilla, ZSTD(1)),
             val_max       SimpleAggregateFunction(max, Float64)  CODEC(Gorilla, ZSTD(1)),
